requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
redis==5.0.1
openai==1.10.0
pydantic==2.5.3
python-multipart==0.0.6
httpx==0.26.0
python-dotenv==1.0.0
llama-index==0.9.39
llama-index-vector-stores-redis==0.1.0
numpy==1.24.3